+++ examples/src/utils/resources.js
@@ -7,7 +7,7 @@ export const vdbList = {};
 export const loadAndCacheVDB = async (name, file) => {
   if (!vdbList[file]) {
     const vdb = await new Promise(resolve => {
-      new OpenVDB.VDBLoader().load('./assets/bunny_cloud.vdb', (vdb) => {
+      new OpenVDB.VDBLoader().load(file, (vdb) => {
         resolve(vdb);  
       }, null, () => {
         alert('Could not load the VDB file.');